+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## main
 ### ✨ Features and improvements
 - _...Add new stuff here..._
+- Add quaternion-based globe panning mode via `DragPanOptions.fixedBearing` ([#5330](https://github.com/maplibre/maplibre-gl-js/pull/5330))
 
 ### 🐞 Bug fixes
 - _...Add new stuff here..._